Fixed tests not working due to MainWindowViewModel loading views
The tests were failing because they couldn't access a static resource that is in App.xaml.
This is because the tests never load App.xaml; they are loading the views via the MainWindowViewModel ctor.

As a temporary fix, what we can do is make sure that, in the unit tests, we avoid creating these views,
but realisitically we should probably move this overlay creation into the overlay controller
(or any other suitable solution that allows us to still unit test the view model on its own)